Refactor : connection 파일 databaseModels에서 분리
상세 : connection.js 파일 databaseFiles로 외부화
이유 : databaseModels에 들어있기에는 파일의 목적이 맞지 않는다고 판단
